Fixes Telegram ID validation and auth scheme
Updates Telegram ID validation range to long.MaxValue to accommodate larger IDs.
Corrects the authentication scheme to "bearer" for proper authorization header recognition in Swagger.
Moves the GetAllByTelegramIdMessagesRequest from the body to the route.

+++ OpenVPNGateMonitor.SharedModels/DataGateMonitorBackend/TelegramBotIncomingMessageLog/Requests/GetAllByTelegramIdMessagesRequest.cs
@@ -6,7 +6,7 @@ namespace OpenVPNGateMonitor.SharedModels.DataGateMonitorBackend.TelegramBotInco
 public class GetAllByTelegramIdMessagesRequest
 {
     [Required(ErrorMessage = "telegramId is required.")]
-    [Range(1, int.MaxValue, ErrorMessage = "telegramId must be greater than 0.")]
+    [Range(1, long.MaxValue, ErrorMessage = "telegramId must be greater than 0.")]
     [FromRoute(Name = "telegramId")]
     public long TelegramId { get; set; }
